+++ iamgenius/app/components/client-side/PrintPDF/index.js
@@ -4,10 +4,10 @@ import styles from './print-pdf.module.css';
 
 import React from 'react';
 import { useReactToPrint } from 'react-to-print';
-import { PDFIcon } from '../server-side/IconLinks';
-import LocalizedMessage from '../client-side/LocalizedMessage';
+import { PDFIcon } from '../../server-side/IconLinks';
+import LocalizedMessage from '../../client-side/LocalizedMessage';
 
-import { slugify } from '../../lib/utils';
+import { slugify } from '../../../lib/utils';
 
 function PrintPDF({ children, width, height, name, language }) {
   const componentRef = React.useRef(null);
